app/requirements.txt
# requirements.txt (CHANGED - removed sentence-transformers and numpy)
fastapi
uvicorn
pymongo
python-dotenv
pydantic